build accepts an arguments which decide on the npm script that will run after this build
